logging. routes. deleting.

+++ Web/Controllers/FeedController.cs
@@ -4,6 +4,7 @@ using DIY_PodcastRss.Utils;
 using DIY_PodcastRss.ViewModels;
 using DIYPodcastRss.Core;
 using DIYPodcastRss.Core.Model;
+using Newtonsoft.Json;
 using System;
 using System.Linq;
 using System.Web.Mvc;
@@ -17,6 +18,7 @@ namespace DIY_PodcastRss.Controllers
             if (CookieHelper.UserUniqueId.IsNullOrWhiteSpace())
             {
                 CookieHelper.UserUniqueId = Guid.NewGuid().ToString();
+                Logger.LogMsg("New user! ", CookieHelper.UserUniqueId, System.Web.HttpContext.Current.Request.ServerVariables["HTTP_X_FORWARDED_FOR"] ?? System.Web.HttpContext.Current.Request.ServerVariables["REMOTE_ADDR"]);
             }
 
         }
@@ -26,17 +28,39 @@ namespace DIY_PodcastRss.Controllers
             return View(vm);
         }
 
+        private string TryMakeDateFromJavaScriptDate()
+        {
+            try
+            {
+                string[] browserDateParts = Request.Form["UserBrowserDateTime"].Split(null);
+                return string.Join(" ", browserDateParts.Take(5));
+            }
+            catch (Exception)
+            {
+                return null;
+            }
+        }
+
+        [Throttle(Name = "CreateFeedThrottle", Seconds = 5)]
         [HttpPost]
         public ActionResult Create(UserFeed postedUserFeed)
         {
+            Logger.LogMsg("Request to create feed ", Environment.NewLine, postedUserFeed.FeedName, postedUserFeed.ImgUrl, string.Join(",", postedUserFeed.Files));
 
             if (ModelState.IsValid)
             {
+                if (postedUserFeed.FeedName.IsNullOrWhiteSpace())
+                {
+                    postedUserFeed.FeedName = TryMakeDateFromJavaScriptDate() ?? DateTime.UtcNow.FriendlyFormat();
+                }
                 postedUserFeed.Files = Request.Form["Files"].ToString().Split(null).Where(x => x.HasValue()).ToList();
                 postedUserFeed.CreatedOnUtc = DateTime.UtcNow;
                 postedUserFeed.BaseUrl = Request.Url.GetLeftPart(UriPartial.Authority) + Url.Content("~");
                 postedUserFeed.CreatedFromIpHost = "{0} {1}".FormatWith(Request.UserHostAddress, Request.UserHostName);
                 postedUserFeed.UserUniqueId = CookieHelper.UserUniqueId;
+
+                Logger.LogMsg("Cleaned up new feed ", Environment.NewLine, JsonConvert.SerializeObject(postedUserFeed, Formatting.Indented));
+
                 var rssGenerator = new DIYPodcastRss.Core.RssGenerator();
                 var syndicationFeed = rssGenerator.CreateRss(postedUserFeed);
                 var feedResult = new SyndicationFeedResult();
@@ -44,22 +68,21 @@ namespace DIY_PodcastRss.Controllers
                 postedUserFeed.FeedDocument = feedResult.GenerateRssXml(syndicationFeed);
                 var repo = new FeedRepo();
 
-
                 //ensure the feed token is unique
-                while (repo.ReserveEmptyFeed(postedUserFeed.FeedToken)==false)
+                while (repo.ReserveEmptyFeed(postedUserFeed.FeedToken) == false)
                 {
-                   postedUserFeed.FeedToken = GuidEncoder.New();
+                    postedUserFeed.FeedToken = GuidEncoder.New();
                 }
 
                 repo.SaveFeed(postedUserFeed);
                 ViewBag.NewFeedToken = postedUserFeed.FeedToken;
-                return View(postedUserFeed);
             }
             return View(postedUserFeed);
         }
 
         public ActionResult All()
         {
+            Logger.LogMsg("All Feeds", CookieHelper.UserUniqueId);
             var vm = new AllUserFeedsViewModel();
             var repo = new FeedRepo();
             vm.Feeds = repo.AllFeeds().OrderByDescending(x => x.CreatedOnUtc);
@@ -68,28 +91,44 @@ namespace DIY_PodcastRss.Controllers
 
         public ActionResult MyFeeds()
         {
+            Logger.LogMsg("My Feed for User ", CookieHelper.UserUniqueId);
             return RedirectToRoute("UserFeeds", new { userId = CookieHelper.UserUniqueId });
         }
 
         public ActionResult UserFeeds(string userId)
         {
+            Logger.LogMsg("User Feed for User ", CookieHelper.UserUniqueId);
             var repo = new FeedRepo();
             var vm = new UserHistoryViewModel();
-            vm.Feeds = repo.AllFeeds().Where(x => x.UserUniqueId == userId && x.DeletedOnUtc.HasValue==false).OrderByDescending(x => x.CreatedOnUtc);
+            vm.Feeds = repo.AllFeeds().Where(x => x.UserUniqueId == userId && x.DeletedOnUtc.HasValue == false).OrderByDescending(x => x.CreatedOnUtc);
+            Logger.LogMsg("Number feeds found:", vm.Feeds.Count());
+
             return View(vm);
         }
 
+        [Throttle(Name = "DeleteFeedThrottle", Seconds = 5)]
+        [HttpPost]
         public ActionResult Delete(string feedToken)
         {
+            Logger.LogMsg("Deleting feed ", feedToken, " User Id ", CookieHelper.UserUniqueId, "at", Request.UserHostAddress);
             var repo = new FeedRepo();
-            bool repoWasDeleted = repo.DeleteFeed(feedToken);
-            ViewBag.SuccessfullyDeletedThatFeed = repoWasDeleted;
-            return RedirectToRoute("MyFeeds");
+            var callingUserId = CookieHelper.UserUniqueId;
+            if (repo.UserCanDeleteFeed(feedToken, callingUserId))
+            {
+                bool repoWasDeleted = repo.DeleteFeed(feedToken);
+                return Json(repoWasDeleted);
+            }
+            else
+            {
+                Response.StatusCode = 500;
+                return null;
+            }
         }
 
 
         public string ViewFeed(string feedToken)
         {
+            Logger.LogMsg("Viewing feed ", feedToken);
             var repo = new FeedRepo();
             var feed = repo.GetFeed(feedToken);
             if (feed != null)
